Added functionality , fixed bug
The functionality of adding a task is implemented. The task is added with a status to do (hardcoded). Task without text won't be added.

A bug which was causing the necessity to re-render App on load to show the current tasks from firebase was fixed.

Some unfinal styling has been added so it could be used as a base when working on styling later.

An Info msg has been added for each container to show when there aren't tasks from that specific status.

The data has an attribute key now that takes the firebase key of a record.