Mark a task as Done. Implemented delete
A task can now be marked as done. When status is done it will give the user a chance to return the task in progress. That will override the assigned date. Later when the task is marked as Done it will override the completed date.

Since it is  in the requirements I decided to implement the opportunity to delete a task.

The timeline of the edits (dates) is now shown on hover of a google image.