Error handling
A proper error or info message is now live. On second thought there would have been better to have one component Message and the class to change based on error, info, success, loading for instance.  The messages are show together with the panel with tasks, which is a different implementation of status than earlier udner the presentation.